refactor: extract gateway tool orchestration
