+++ crates/agentic-core/src/executor/engine.rs
@@ -9,29 +9,27 @@ use std::sync::Arc;
 
 use async_stream::stream;
 use either::Either;
-use futures::StreamExt;
-use futures::stream as futures_stream;
-use serde_json::Value;
 use tokio::sync::mpsc;
 use tracing::warn;
 
+use super::gateway::{
+    append_gateway_calls_to_new_input, append_output_items_to_input, append_tool_outputs,
+    execute_and_emit_output_calls, has_client_owned_calls, public_output_items,
+};
 use crate::executor::accumulator::ResponseAccumulator;
 use crate::executor::error::{ExecutorError, ExecutorResult};
 use crate::executor::inference::{DONE_MARKER, call_inference, fetch_response_json};
 use crate::executor::persist::persist_response;
 use crate::executor::rehydrate::rehydrate_conversation;
 use crate::executor::request::{ExecutionContext, RequestContext};
-use crate::tool::{ToolError, ToolOutput, ToolRegistry, ToolType};
-use crate::types::io::output::{FunctionToolCall, WebSearchCall, WebSearchCallStatus, WebSearchSource};
-use crate::types::io::{InputItem, OutputItem, ResponseUsage, ResponsesInput, ToolChoice};
+use crate::tool::ToolRegistry;
+use crate::types::io::{ResponseUsage, ToolChoice};
 use crate::types::request_response::{RequestPayload, ResponsePayload};
 use crate::utils::common::serialize_to_string;
 
 pub use crate::executor::inference::BoxStream;
 
 const MAX_GATEWAY_TOOL_ROUNDS: usize = 10;
-const MAX_GATEWAY_TOOL_CALLS_PER_ROUND: usize = 8;
-const MAX_GATEWAY_TOOL_CONCURRENCY: usize = 4;
 
 fn should_persist(ctx: &RequestContext) -> bool {
     ctx.original_request.store
@@ -87,82 +85,6 @@ async fn fetch_stream_payload(
     Ok(payload)
 }
 
-fn append_input_item(input: &mut ResponsesInput, item: InputItem) {
-    match input {
-        ResponsesInput::Items(items) => items.push(item),
-        ResponsesInput::Text(text) => {
-            let text_input = ResponsesInput::Text(std::mem::take(text));
-            let mut items = Vec::<InputItem>::from(&text_input);
-            items.push(item);
-            *input = ResponsesInput::Items(items);
-        }
-    }
-}
-
-fn append_output_items_to_input(input: &mut ResponsesInput, output_items: &[OutputItem]) {
-    for output_item in output_items {
-        let input_item = match output_item {
-            OutputItem::Message(message) => Some(InputItem::Message(message.clone().into())),
-            OutputItem::FunctionCall(call) => Some(InputItem::FunctionCall(call.clone())),
-            OutputItem::Reasoning(reasoning) => Some(InputItem::Reasoning(reasoning.clone())),
-            OutputItem::WebSearchCall(_) | OutputItem::Unknown => None,
-        };
-        if let Some(input_item) = input_item {
-            append_input_item(input, input_item);
-        }
-    }
-}
-
-fn append_tool_outputs(ctx: &mut RequestContext, tool_outputs: Vec<InputItem>) {
-    for output in tool_outputs {
-        ctx.new_input_items.push(output.clone());
-        append_input_item(&mut ctx.enriched_request.input, output);
-    }
-}
-
-fn function_calls(output_items: &[OutputItem]) -> Vec<FunctionToolCall> {
-    output_items
-        .iter()
-        .filter_map(|item| match item {
-            OutputItem::FunctionCall(call) => Some(call.clone()),
-            _ => None,
-        })
-        .collect()
-}
-
-fn is_gateway_owned_call(call: &FunctionToolCall, registry: &ToolRegistry) -> bool {
-    registry
-        .lookup(&call.name)
-        .is_some_and(|entry| entry.tool_type != ToolType::Function)
-}
-
-fn append_gateway_calls_to_new_input(ctx: &mut RequestContext, output_items: &[OutputItem], registry: &ToolRegistry) {
-    ctx.new_input_items.extend(output_items.iter().filter_map(|item| {
-        let OutputItem::FunctionCall(call) = item else {
-            return None;
-        };
-        is_gateway_owned_call(call, registry).then(|| InputItem::FunctionCall(call.clone()))
-    }));
-}
-
-fn public_output_items(
-    output_items: Vec<OutputItem>,
-    registry: &ToolRegistry,
-    gateway_results: &[GatewayCallResult],
-) -> Vec<OutputItem> {
-    output_items
-        .into_iter()
-        .map(|item| match item {
-            OutputItem::FunctionCall(call) if is_gateway_owned_call(&call, registry) => gateway_results
-                .iter()
-                .find(|result| result.call.call_id == call.call_id)
-                .and_then(|result| result.public_output.clone())
-                .unwrap_or(OutputItem::FunctionCall(call)),
-            other => other,
-        })
-        .collect()
-}
-
 fn add_usage(total: ResponseUsage, usage: ResponseUsage) -> ResponseUsage {
     ResponseUsage {
         input_tokens: total.input_tokens.saturating_add(usage.input_tokens),
@@ -187,240 +109,6 @@ fn accumulate_usage(total: &mut Option<ResponseUsage>, usage: Option<ResponseUsa
     if let Some(usage) = usage {
         *total = Some(total.map_or(usage, |current| add_usage(current, usage)));
     }
-}
-
-struct GatewayCallDispatch {
-    call: FunctionToolCall,
-    tool_type: ToolType,
-    config: Value,
-    executor: Arc<dyn crate::tool::GatewayExecutor>,
-}
-
-struct GatewayCallExecution {
-    call: FunctionToolCall,
-    tool_type: ToolType,
-    output: Result<ToolOutput, ToolError>,
-}
-
-#[derive(Clone)]
-struct GatewayCallResult {
-    call: FunctionToolCall,
-    input_item: InputItem,
-    public_output: Option<OutputItem>,
-}
-
-struct GatewayCallEventPlan {
-    call_id: String,
-    output_index: u32,
-    started_output: Option<OutputItem>,
-}
-
-fn gateway_dispatches(
-    calls: &[FunctionToolCall],
-    registry: &ToolRegistry,
-    exec_ctx: &ExecutionContext,
-) -> ExecutorResult<Vec<GatewayCallDispatch>> {
-    let dispatches = registry
-        .gateway_owned(calls)
-        .into_iter()
-        .map(|call| {
-            let entry = registry.lookup(&call.name).ok_or_else(|| {
-                ToolError::Config(format!(
-                    "gateway tool '{}' was not found in the request registry",
-                    call.name
-                ))
-            })?;
-            let executor = exec_ctx
-                .gateway_executors
-                .get(entry.tool_type)
-                .ok_or_else(|| ToolError::Config(format!("no gateway executor registered for tool '{}'", call.name)))?;
-            Ok(GatewayCallDispatch {
-                call: call.clone(),
-                tool_type: entry.tool_type,
-                config: entry.config.clone(),
-                executor,
-            })
-        })
-        .collect::<Result<Vec<_>, ToolError>>()
-        .map_err(ExecutorError::from)?;
-    if dispatches.len() > MAX_GATEWAY_TOOL_CALLS_PER_ROUND {
-        return Err(ExecutorError::InvalidRequest(format!(
-            "gateway tool call limit exceeded: got {}, max {MAX_GATEWAY_TOOL_CALLS_PER_ROUND} per round",
-            dispatches.len()
-        )));
-    }
-    Ok(dispatches)
-}
-
-async fn execute_gateway_dispatch(dispatch: GatewayCallDispatch) -> GatewayCallExecution {
-    let GatewayCallDispatch {
-        call,
-        tool_type,
-        config,
-        executor,
-    } = dispatch;
-    let output = executor
-        .execute(&call.call_id, &call.name, &call.arguments, &config)
-        .await;
-    GatewayCallExecution {
-        call,
-        tool_type,
-        output,
-    }
-}
-
-fn execution_error_output(call: &FunctionToolCall, message: &str) -> ExecutorResult<ToolOutput> {
-    let output = serialize_to_string(&serde_json::json!({ "error": message })).map_err(ExecutorError::JsonError)?;
-    Ok(ToolOutput {
-        call_id: call.call_id.clone(),
-        output,
-    })
-}
-
-fn gateway_public_output(
-    tool_type: ToolType,
-    call: &FunctionToolCall,
-    output: &ToolOutput,
-    status: WebSearchCallStatus,
-) -> Option<OutputItem> {
-    match tool_type {
-        ToolType::WebSearch => Some(web_search_output_item(call, output, status)),
-        ToolType::Function | ToolType::Mcp | ToolType::FileSearch | ToolType::CodeInterpreter => None,
-    }
-}
-
-async fn execute_gateway_calls(
-    calls: &[FunctionToolCall],
-    registry: &ToolRegistry,
-    exec_ctx: &ExecutionContext,
-) -> ExecutorResult<Vec<GatewayCallResult>> {
-    let dispatches = gateway_dispatches(calls, registry, exec_ctx)?;
-    let executions = futures_stream::iter(dispatches.into_iter().map(execute_gateway_dispatch))
-        .buffered(MAX_GATEWAY_TOOL_CONCURRENCY)
-        .collect::<Vec<_>>()
-        .await;
-    let mut results = Vec::with_capacity(executions.len());
-
-    for execution in executions {
-        let GatewayCallExecution {
-            call,
-            tool_type,
-            output,
-        } = execution;
-        let (output, status) = match output {
-            Ok(output) => (output, WebSearchCallStatus::Completed),
-            Err(ToolError::Execution(message)) => {
-                (execution_error_output(&call, &message)?, WebSearchCallStatus::Failed)
-            }
-            Err(ToolError::Config(message)) if tool_type == ToolType::WebSearch => {
-                (execution_error_output(&call, &message)?, WebSearchCallStatus::Failed)
-            }
-            Err(err @ ToolError::Config(_)) => return Err(err.into()),
-        };
-        let public_output = gateway_public_output(tool_type, &call, &output, status);
-        results.push(GatewayCallResult {
-            call,
-            input_item: InputItem::FunctionCallOutput(output.into()),
-            public_output,
-        });
-    }
-
-    Ok(results)
-}
-
-fn web_search_output_item(call: &FunctionToolCall, output: &ToolOutput, status: WebSearchCallStatus) -> OutputItem {
-    let parsed_output = serde_json::from_str::<Value>(&output.output).ok();
-    let query = parsed_output
-        .as_ref()
-        .and_then(|value| clean_json_str(value.get("query")))
-        .or_else(|| web_search_query_from_arguments(&call.arguments))
-        .unwrap_or_default();
-    let sources = parsed_output
-        .as_ref()
-        .map(web_search_sources_from_output)
-        .unwrap_or_default();
-    OutputItem::WebSearchCall(WebSearchCall::new(web_search_call_id(call), status, query, sources))
-}
-
-fn started_web_search_output_item(call: &FunctionToolCall) -> OutputItem {
-    OutputItem::WebSearchCall(WebSearchCall::new(
-        web_search_call_id(call),
-        WebSearchCallStatus::InProgress,
-        web_search_query_from_arguments(&call.arguments).unwrap_or_default(),
-        Vec::new(),
-    ))
-}
-
-fn web_search_call_id(call: &FunctionToolCall) -> String {
-    if let Some(suffix) = call.id.strip_prefix("fc_").filter(|suffix| !suffix.is_empty()) {
-        return format!("ws_{suffix}");
-    }
-    if let Some(suffix) = call.call_id.strip_prefix("call_").filter(|suffix| !suffix.is_empty()) {
-        return format!("ws_{suffix}");
-    }
-    crate::utils::uuid7_str("ws_")
-}
-
-fn web_search_query_from_arguments(arguments: &str) -> Option<String> {
-    let args = serde_json::from_str::<Value>(arguments).ok()?;
-    clean_json_str(args.get("query"))
-}
-
-fn web_search_sources_from_output(output: &Value) -> Vec<WebSearchSource> {
-    ["web", "news"]
-        .into_iter()
-        .filter_map(|section| output.get("results")?.get(section)?.as_array())
-        .flat_map(|results| results.iter())
-        .filter_map(web_search_source_from_result)
-        .collect()
-}
-
-fn web_search_source_from_result(result: &Value) -> Option<WebSearchSource> {
-    let url = clean_json_str(result.get("url"))?;
-    Some(WebSearchSource {
-        url,
-        title: clean_json_str(result.get("title")),
-    })
-}
-
-fn clean_json_str(value: Option<&Value>) -> Option<String> {
-    value
-        .and_then(Value::as_str)
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(str::to_owned)
-}
-
-fn gateway_event_plans(
-    output_items: &[OutputItem],
-    registry: &ToolRegistry,
-    output_offset: usize,
-) -> Vec<GatewayCallEventPlan> {
-    let mut output_index = output_offset;
-    let mut plans = Vec::new();
-    for item in output_items {
-        if let OutputItem::FunctionCall(call) = item
-            && let Some(entry) = registry.lookup(&call.name)
-            && entry.tool_type != ToolType::Function
-        {
-            plans.push(GatewayCallEventPlan {
-                call_id: call.call_id.clone(),
-                output_index: u32::try_from(output_index).unwrap_or(u32::MAX),
-                started_output: match entry.tool_type {
-                    ToolType::WebSearch => Some(started_web_search_output_item(call)),
-                    ToolType::Function | ToolType::Mcp | ToolType::FileSearch | ToolType::CodeInterpreter => None,
-                },
-            });
-        }
-        output_index = output_index.saturating_add(1);
-    }
-    plans
-}
-
-fn emit_sse_json(sender: &mpsc::UnboundedSender<String>, event: &Value) -> ExecutorResult<()> {
-    let event_json = serialize_to_string(&event).map_err(ExecutorError::JsonError)?;
-    let _ = sender.send(format!("data: {event_json}\n\n"));
-    Ok(())
 }
 
 fn error_sse_chunk(message: &str) -> String {
@@ -461,82 +149,6 @@ impl<T> Drop for AbortOnDrop<T> {
     }
 }
 
-fn output_item_value(item: &OutputItem) -> ExecutorResult<Value> {
-    serde_json::to_value(item).map_err(ExecutorError::JsonError)
-}
-
-fn emit_gateway_start_events(
-    plans: &[GatewayCallEventPlan],
-    stream_events: Option<&mpsc::UnboundedSender<String>>,
-) -> ExecutorResult<()> {
-    let Some(sender) = stream_events else {
-        return Ok(());
-    };
-    for plan in plans {
-        let Some(output_item) = &plan.started_output else {
-            continue;
-        };
-        let OutputItem::WebSearchCall(web_search_call) = output_item else {
-            continue;
-        };
-        let item = output_item_value(output_item)?;
-        let added_event = serde_json::json!({
-                "type": "response.output_item.added",
-                "output_index": plan.output_index,
-                "item": item
-        });
-        emit_sse_json(sender, &added_event)?;
-        let in_progress_event = serde_json::json!({
-                "type": "response.web_search_call.in_progress",
-                "item_id": web_search_call.id,
-                "output_index": plan.output_index
-        });
-        emit_sse_json(sender, &in_progress_event)?;
-        let searching_event = serde_json::json!({
-                "type": "response.web_search_call.searching",
-                "item_id": web_search_call.id,
-                "output_index": plan.output_index
-        });
-        emit_sse_json(sender, &searching_event)?;
-    }
-    Ok(())
-}
-
-fn emit_gateway_completed_events(
-    results: &[GatewayCallResult],
-    plans: &[GatewayCallEventPlan],
-    stream_events: Option<&mpsc::UnboundedSender<String>>,
-) -> ExecutorResult<()> {
-    let Some(sender) = stream_events else {
-        return Ok(());
-    };
-    for result in results {
-        let Some(OutputItem::WebSearchCall(web_search_call)) = &result.public_output else {
-            continue;
-        };
-        let output_index = plans
-            .iter()
-            .find(|plan| plan.call_id == result.call.call_id)
-            .map_or(0, |plan| plan.output_index);
-        let output_item = OutputItem::WebSearchCall(web_search_call.clone());
-        let item = output_item_value(&output_item)?;
-        let completed_event = serde_json::json!({
-                "type": "response.web_search_call.completed",
-                "item_id": web_search_call.id,
-                "output_index": output_index,
-                "item": item.clone()
-        });
-        emit_sse_json(sender, &completed_event)?;
-        let done_event = serde_json::json!({
-                "type": "response.output_item.done",
-                "output_index": output_index,
-                "item": item
-        });
-        emit_sse_json(sender, &done_event)?;
-    }
-    Ok(())
-}
-
 async fn run_until_gateway_tools_complete(
     mut ctx: RequestContext,
     exec_ctx: &ExecutionContext,
@@ -548,7 +160,9 @@ async fn run_until_gateway_tools_complete(
         .enriched_request
         .tools
         .as_ref()
-        .map_or_else(ToolRegistry::default, |tools| ToolRegistry::build(tools));
+        .map_or_else(ToolRegistry::default, |tools| {
+            ToolRegistry::build_with_handlers(tools, |tool_type| exec_ctx.gateway_executors.get(tool_type))
+        });
     let mut combined_output = Vec::new();
     let mut combined_usage = None;
 
@@ -560,13 +174,10 @@ async fn run_until_gateway_tools_complete(
         };
         accumulate_usage(&mut combined_usage, payload.usage);
         let current_output = std::mem::take(&mut payload.output);
-        let calls = function_calls(&current_output);
-        let has_client_owned_calls = !registry.client_owned(&calls).is_empty();
-        let event_plans = gateway_event_plans(&current_output, &registry, combined_output.len());
-        emit_gateway_start_events(&event_plans, stream_events)?;
-        let gateway_results = execute_gateway_calls(&calls, &registry, exec_ctx).await?;
-        emit_gateway_completed_events(&gateway_results, &event_plans, stream_events)?;
-        let public_output = public_output_items(current_output.clone(), &registry, &gateway_results);
+        let has_client_owned_calls = has_client_owned_calls(&current_output, &registry);
+        let gateway_results =
+            execute_and_emit_output_calls(&current_output, &registry, combined_output.len(), stream_events).await?;
+        let public_output = public_output_items(&current_output, &registry, &gateway_results);
 
         if has_client_owned_calls {
             combined_output.extend(public_output);
