+++ crates/agentic-core/src/executor/mod.rs
@@ -9,6 +9,8 @@ pub mod persist;
 pub mod rehydrate;
 pub mod request;
 
+mod gateway;
+
 pub use engine::{BoxStream, ExecuteRequest, create_conversation, execute};
 pub use error::{ExecutorError, ExecutorResult};
 pub use inference::call_inference;
