+++ crates/agentic-core/src/tool/mod.rs
@@ -11,5 +11,5 @@ pub mod web_search;
 
 pub use function::FunctionHandler;
 pub use handler::{GatewayExecutor, ToolError, ToolHandler, ToolOutput};
-pub use registry::{ToolEntry, ToolRegistry, ToolType};
+pub use registry::{GatewayDispatchResult, ToolEntry, ToolRegistry, ToolType};
 pub use web_search::WebSearchHandler;
